feat(TiszaBuilder): Update Utils round
- Change factor type to double for improved accuracy.

+++ TiszaBuilder/src/main/java/hu/ikoli/tiszaBuilder/utils/Utils.java
@@ -37,10 +37,10 @@ public class Utils {
         if (places < 0)
             throw new IllegalArgumentException();
 
-        long factor = (long) Math.pow(10, places);
+        double factor = (long) Math.pow(10, places);
         value = value * factor;
-        long tmp = Math.round(value);
-        return (double) tmp / factor;
+        double tmp = Math.round(value);
+        return tmp / factor;
     }
 
 }
